+++ src/tests/simple/Audio2Test/buffer/IAudioSampleProvider.cpp
@@ -8,16 +8,16 @@
 #include <cassert>
 #include <cmath>
 
-static inline void boundCheck(const IAudioSampleProvider &iAudioStorage, int channel, int startPos, int length) {
+static inline void boundCheck(const IAudioSampleProvider &iAudioStorage, int channel, qint64 startPos, qint64 length) {
     assert(channel >= 0 && channel < iAudioStorage.channelCount());
     assert(startPos >= 0 && startPos <= iAudioStorage.sampleCount());
     assert(startPos + length >= 0 && startPos + length <= iAudioStorage.sampleCount());
 }
 
-float IAudioSampleProvider::magnitude(int channel, int startPos, int length) const {
+float IAudioSampleProvider::magnitude(int channel, qint64 startPos, qint64 length) const {
     boundCheck(*this, channel, startPos, length);
     float m = 0;
-    for(int i = 0; i < length; i++) {
+    for(qint64 i = 0; i < length; i++) {
         m = std::max(m, abs(constSampleAt(channel, startPos + i)));
     }
     return m;
@@ -25,10 +25,10 @@ float IAudioSampleProvider::magnitude(int channel, int startPos, int length) con
 float IAudioSampleProvider::magnitude(int channel) const {
     return magnitude(channel, 0, sampleCount());
 }
-float IAudioSampleProvider::rms(int channel, int startPos, int length) const {
+float IAudioSampleProvider::rms(int channel, qint64 startPos, qint64 length) const {
     boundCheck(*this, channel, startPos, length);
     float s = 0;
-    for(int i = 0; i < length; i++) {
+    for(qint64 i = 0; i < length; i++) {
         auto sample = constSampleAt(channel, startPos + i);
         s += sample * sample;
     }
